Fix incorrect setting of m_full_key_lookup
Summary:
This fixes https://github.com/facebook/mysql-5.6/issues/1174

For secondary keys, it is not sufficient to look at keypart_map to determine whether the full is passed in or not, since the sql layer sometimes passes in `HA_WHOLE_KEY` even when it does not include extended keys.

We already have handling for it, and we only do point lookups when the secondary key includes extended keys as well. However, we were always setting `m_full_key_lookup` even if we end up not doing a point lookup.

Reviewed By: yoshinorim

Differential Revision: D31862865

fbshipit-source-id: be77d1199f6