Add force-index option for mysqldump
Summary:
For backing up MyRocks databases, the preference is to order by
the primary key since bulk loading on the restore side will ensure the
fastest load times.

However, tables whose primary keys consist of prefix fields may run
into performance problems during the dump because mysqldump will try to
select them with an order by clause. Since the prefix field is not the
entire field, mysqld will attempt to filesort the data, leading to a
slow dump phase.

--order-by-primary may also return data that can cause bulk loading to fail in the prefix key case.

Since the data in the primary key is already sorted in primary key
order, a workaround would be to perform the select * with a force index
hint if the table is either InnoDB or RocksDB, and skip the order by
clause.

Reviewed By: Pushapgl, yizhang82

Differential Revision: D31976467