Fixing race between STOP SLAVE and SQL thread error handling
Summary:
When there is an error in a worker thread, it sets the killed
flag on the SQL thread to KILL_QUERY before terminating. In the SQL
thread exit flow we use the killed flag to figure out if we have an
error situation, if yes, we skip waiting for workers to finish executing
in flight transactions (since they might not be able to finish execution
due to another worker erroring out). STOP SLAVE command sets the killed
flag of SQL thread to NOT_KILLED in terminate_slave_thread() method. So,
if STOP SLAVE races with SQL thread exit flow, we might end up with
killed = NOT_KILLED just before deciding if we should be waiting for
workers to finish execution. Which can lead to SQL thread waiting for
worker indefinitely.

To fix the issue we set a another flag when worker threads exit due to
an error and check this flag in SQL thread along with the killed flag.

Reviewed By: hermanlee

Differential Revision: D32511539

fbshipit-source-id: 456375fd37b