Fixed flaky esms_by_all test
Summary:
Sample query is getting refreshed by new_max_wait which is set by comparing the wait_time stored in pfs struct. As of today there is no good way to control that behavior, so removing the queries which are creating same digest.

We already have other tests which are checking the code when the query is refreshed after wait_time, so there is no need to test that code as part of this test.

Related code which is updating new_max_wait.

```
 /* Update the digest sample if it's a new maximum. */
      if (wait_time > digest_stat->get_sample_timer_wait()) {
        new_max_wait = true;
      }
```
```
if (!get_sample_query) {
        get_sample_query = new_max_wait;

        if (!get_sample_query) {
          /* Check the query sample age. */
          if (pfs_param.m_max_digest_sample_age > 0) {
            /* Comparison in micro seconds. */
            get_sample_query = (digest_stat->get_sample_age() >
                                pfs_param.m_max_digest_sample_age * 1000000);
          }
        }
      }
```

Reviewed By: george-reynya

Differential Revision: D32572018

fbshipit-source-id: a3ef81873e9