Bug#33300271: EXPLAIN + CTE + temporary table leads to crash mysqld
Summary:
During resolving, for an always true or an always false condition, we
remove all items that are part of the condition and replace with true
or false items. If the removed condition has a subquery that refers
to a common table expression, we cannot remove such a subquery without
removing the reference stored in the common table expression to the
table in the subquery. Otherwise we fail trying to access the
reference after the subquery's removal.
Fix it by removing the reference in the common table expression when
the query block for the subquery gets removed.

Change-Id: I33208e19b9e0e351c1d1dfbb884a4f7e7e264a56

Reviewed By: george-reynya

Differential Revision: D45163386

fbshipit-source-id: 0f22e877759e3837c5d39a68347bdf9856de9f59