Calling after_commit in commit flows where ordered_commit() is not called
Summary:
During crash recovery if engine gtid > last raft log gtid i.e.
when tail of raft log is lost due to infrequent fsync, the applier
applies all transactions from the tip of the truncated raft log to
engine gtid in idempotent mode. The LWM applied marker in the plugin is
bootstrapped with the last opid in the raft log which is also < engine
gtid.

Because these trxs are already included in executed gtids, row event
execution is skipped which leads to binlog cache not being initialized.
When binlog cache is not initialized MYSQL_BIN_LOG::ordered_commit() is
not called, instead ha_commit_low() is called directly in
MYSQL_BIN_LOG::commit(). Not calling ordered_commit() means skipping
after_commit hook for these trxs which causes gaps in the applied trxs
on the plugin side.

To fix the issue setting a flag in thd when row event execution is
skipped and calling after_commit in ha_commit_low when this flag is set.

This is a port of D32606617 (https://github.com/facebook/mysql-5.6/commit/069df5730e8675ec4b292cc7d8430a3290a17699)

Reviewed By: bhatvinay

Differential Revision: D32654212

fbshipit-source-id: d93f2ed90d1