failure injection framework
Summary:
Initial implementation of a framework which supports injecting failure in regular builds. This will be useful to simulate different scenarios and failures in production environment.

A new failure injection class maintains all activation for 'failure-points'. All failure points are defined through a enum 'Failure_points'. An example failure point is implemented which induces a delay/stall on every binlog rotation.

Failure points can take values which can be of special meaning to the way failure is injected (to be implemented by each failure injecting function). The framework provides a mechanism for the failure injection code to get the value associated with the failure point as a string representation.

For example, STALL_BINLOG_ROTATE is a failure point which can take an optional value for how long the stall should be (in milli-seconds).

There are two sysvars which control how the feature is used.
A read-only sysvar 'enable_failure_injection' is used to turn ON/OFF the entire framework. The value can be modified only through a server restart (as an additional safety mechanism).

A string sysvar 'failure_injection_points' is used to activate specific failure points. It can take a comma seperated list of failure points (and its associated values, if any). For example 'SET GLOBAL failure_injection_points = "inject_binlog_stall=10,inject_write_failure_rate=5,inject_crash_on_election"' activates three failure points (two of which take values).

All the failure points that are activated currently can be queried through a information_schema table 'failure_injection_points'. This simply iterates through the in-memory list of activated failure points (and is not a persistent table)

Future enhancements include adding more failure points and implemeting the failure injection logic for these failure points.

Reviewed By: li-chi

Differential Revision: D32611746

fbshipit-source-id: 439eb1e3f7e