Setting relay_log_basename and relay_log_index correctly when raft is enabled
Summary:
These variables are set in init_server_components() but that
method is only called during server startup. We need to set (and reset)
these variables when raft is enabled/disabled.

This issue was causing transaction retires on secondaries to fail
because to retry transactions we were trying to read the file derived
from the default value of relay_log_basename (mysqld-relay-log) when
raft was enabled and file derived from raft relay log name (binary-log)
after raft was disabled. Restarting after enable/disable of raft solved
the issue (because of init_server_components()) but we should be setting
these vars correctly when raft is enabled/disabled.

To fix the issue setting these vars correctly when enable_raft_plugin
var is updated.

Reviewed By: anirbanr-fb

Differential Revision: D32856678

fbshipit-source-id: ade0d21fa14