Defer AC readmission errors with killed flag
Summary:
During readmission in case of timeout or abort the error is set on `THD` object but is not returned as an error code to the caller of `thd_wait_end()` (which returns void). Turns out this is not a supported scenario leading to some code ignoring the error while some code honoring it and skipping functionality but returning success due to a bug. As a result it is possible that `insert` could end up inserting the rows and committing local transaction but not logging it in the binlog so it's not replicated.

The fix is utilize the killed flag and also store the AC error as "deferred" error on `THD`. When at the end of the statement if the `KILL_QUERY` flag is set and deferred error is found then that error is returned. This approach avoids inconsistencies caused by dealing with "unexpected" error.

Reviewed By: lth

Differential Revision: D33777903

fbshipit-source-id: e618e3f6efd