[rpl_lag_manager] Add logging to mysql instance logs for easier analysis
Summary:
Now we will be able to trace back how write throttling reacted to a case of replication lag on a per issue occurrence basis.

Added relevant logs at a few different places in the logic. Made sure that the logs are added in a way that they do not end up flooding the mysql error logs.

Differential Revision: D34122818