Fix error message in SELECT FOR UPDATE on a read-only instance to report primary info
Summary: It doesn't report new primary information (unlike regular inserts, etc). This is needed for upstream customers to redirect to new primary if SELECT FOR UPDATE is failing. 8.0 doesn't have this problem because the check and error is in a single function. Given that we are moving to 8.0 soon I'm going with a small targeted fix.

Reviewed By: luqun

Differential Revision: D34155382

fbshipit-source-id: 7d0e7445478