Add logging to mysql instance logs for easier analysis
Summary:
This is backport of diff D34122818 from 8.0 to 5.6

Now we will be able to trace back how write throttling reacted to a case of replication lag on a per issue occurrence basis.
Added relevant logs at a few different places in the logic. Made sure that the logs are added in a way that they do not end up flooding the mysql error logs.

Reviewed By: mzait

Differential Revision: D34184860

fbshipit-source-id: 7d5e3790b29