Prevent InnoDB lob::purge() from generating conflicting updates
Summary:
https://bugs.mysql.com/bug.php?id=105592. No fix is yet
available up to 8.0.28, but this can be dropped once this bug is
resolved.

lob::purge() can have conflicting updates to the same page between the
lob_mtr and btr_mtr. These almost always involve pages 0 and 2, the
FSP_HDR and INODE pages. The end result is out of order mtr being
applied during crash recovery, resulting in a corrupted instance.

Work around is to move the remainder of the purge into a separte lob_mtr
too.

Reviewed By: george-reynya

Differential Revision: D34346195

fbshipit-source-id: cfe23d65bd4