Add missing wait for sleep in admission_control_wait_events_errors test
Summary: It is possible that tables are unlocked before con1 gets a chance to start the sleep and claim the only AC slot, so the insert query comes through first without error. The fix is to ensure that sleep is running before unlocking. Actually I saw this issue earlier and fixed it in all test cases in this test but somehow I missed the first one.

Reviewed By: hermanlee

Differential Revision: D34403897

fbshipit-source-id: adf097444ef