Fix in commit order manager when the seq number wraps around
Summary:
This is a port of upstream fix for BUG#32891221:
https://github.com/mysql/mysql-server/commit/f6bb5e7cc5e

In Commit_order_manager::finish_one() a variable was defined using auto which
defaults to int. This causes issues with commit order queues seq number which is
defined as unsigned long long. Also the wrapping logic was not considering 0 and
1 as special seq nums that defines "no seq num" and "frozen seq num"
respectively, this is also handled in this patch by skipping over these values.

NOTE: I did not port over the tests from the original patch since they have a
lot of dependencies. I rewrote the logic of the test in the patch in a new test
instead.

Porting notes: can be dropped when rebasing to 8.0.28.

Reviewed By: hermanlee

Differential Revision: D34554845

fbshipit-source-id: 45f2a5cbc1f