Take snapshots at periodic intervals - initial framework
Summary:
Background:

The SQL statistics are captured today at the end of the execution of a SQL statement. This has two limitations:
a) we do not have details of resource usage of long running queries until they are done
b) the capture into hive is misleading: a long running SQL statement shows in the top of a capture period at the end even though it did not contribute during that time window

With this change, we want to take snapshots of mt stats at periodic intervals in 8.0. We had a 5.6 change: https://www.internalfb.com/diff/D31601612 (https://github.com/facebook/mysql-5.6/commit/7f017fe62725c238618227768d381c2c0fa4e1cd) but 8.0 totally differs in its implementation.

Breaking the change-set into 2 parts for ease of review. The new snapshot interface is only added as part of V2 which I am not consuming right now.

Differential Revision: D34596583

fbshipit-source-id: 9979989dd7bb6761caffc2ee293da2effcd3990e