Registering m_position for 80 for plugin side to keep upto date.
Summary:
Between 5.6 and 8.0 the maintenance of the file write position
and binlog file size, while it is being written has changed. Since
Binlog_ofile is where the m_position pointer is and instead of
my_b_tell, the size of the file is being maintained in m_position
(most likely due to encryption feature), the existing solution of not
keeping m_position up to date and depending on atomic_binlog_end_pos had
several holes. One of the gaps was during binlog truncation when the
raft plugin trims the file and reduces its size, but Rpl_applier_reader
is still pegged at the original size, because update_binlog_end_pos
was called during Dump_log repointing.

The solution to these issues and having a more solid design is that
raft plugin should keep the m_position up to date just like non-raft
code does during "::write" and truncation. This diff sends the position
field as a pointer to plugin which keeps it up to date. The position
filed is paired with the IO_CACHE member of ofile and therefore this
is safe to do, since IO_CACHE is already getting
registered/re-registered at each rotation and promotion.

Reviewed By: li-chi

Differential Revision: D34826266