Fix crash on show processlist + srv_session
Summary:
When processlist iterates the list of THDs, it
accesses the protocol field from those THDs.

For the regular path, this field is generally static,
but for srv_sessions, the protocol is pushed and popped
on each statement. This can allow the `show processlist`
thread to access a protocol object as it is being freed,
causing wild ptr reads

Fix is for srv_session to use the thd_protocol lock, and
have all callers from sql_show.cc use the locks when
accessing a protocol from another thread.

Reviewed By: lth

Differential Revision: D35093058

fbshipit-source-id: dfb42fa7c541e0a1aecdb77a406104e8495cb586