Remove the LD_PRELOAD for libtirpc in safe_process to unblock asandebug testing
Summary: In my-test-run.pl, we use safe_process.pm / safe_process.cc to launch process and as part of that it'll inject tirpc and libcom_err into the process using LD_PRELOAD for some "incompatibility" between libasan and libtirpc. However, LD_PRELOAD will inject platform010 version of tirpc and non-platform010 version of libcom_err leading to load issues, and even if we switch to platform010 it probably won't work as well because it launches /bin/sh with LD_PRELOAD and /bin/sh isn't built with platform010. FWIW I don't think anything that we care about depends on libtirpc so removing LD_PRELOAD for now.

Reviewed By: george-reynya

Differential Revision: D35331678

fbshipit-source-id: b75b5e92a0fed0c336490af46eb3500a37b5a9f2