Fix the apply log variable
Summary:
Somewhere between 8.0.23 and 8.0.20 D22181424 wasn't ported correctly.

For some reason, https://fburl.com/phabricator/zn0seeak added the ptr syntax and looked like it tried to set the ptr vars onto the real vars once on init.

So basically the actual variable used by the server to do logging was
probably populated right, so Raft kept logging, but the variable that
shows up on SQL queries is unpopulated.

So when xtrabackup queries for the file name, it gets an empty value and
errors out.

Reviewed By: li-chi

Differential Revision: D35634384

fbshipit-source-id: dee350ced07ff0036bc2cc9b464e5efe02fd32fc