Fix the apply log variable
Summary:
Somewhere between 8.0.23 and 8.0.20 D22181424 wasn't ported correctly.

https://fburl.com/phabricator/zn0seeak added the ptr syntax.
The issue was that we're freeing a system variable in D22181424. The patch in
https://fburl.com/phabricator/zn0seeak avoiding freeing the system var but ended
up essentially stepsiding the var altogether when the variable was queried.

So basically the actual variable used by the server to do logging was
probably populated right, so Raft kept logging, but the variable that
shows up on SQL queries is unpopulated.

Instead, add a flag to indicate if we strduped into the ptr, and only
free it if so.

Squash with D22181424

Reviewed By: hermanlee

Differential Revision: D38883503

fbshipit-source-id: 4673c07c60e03168d3ba9db7c04fb580f7f26660