Defer cleanup of index statistics to when THD goes out of scope
Summary:
This diff does the following:
1. Defers the cleanup of THD local index stats until THD goes out of scope.
2. Skips aggregation for the stored procedure code path.

This is a medium term fix while I test the long term fix which involves creating custom destructor for RowIterators. It would guarantee there are no race conditions. The long term fix would need a wider code change and we would need to evaluate the performance impact thoroughly as well.

Reviewed By: satya-valluri

Differential Revision: D35855533

fbshipit-source-id: 79af119c6535dbe827df033238d5c9c2eed644ae