D35758321 for 5.6, for binlog name race
Summary:
5.6 is slightly different in that it uses this hot_log flag when
deciding to take locks. However, I noticed it still tries to pull out
the log name outside of the lock, and that could cause races.

Move to take the name after the lock. I believe this is safe since we
don't do anything with the name previously till we take a lock anyway.

Reviewed By: anirbanr-fb

Differential Revision: D36558928

fbshipit-source-id: 47db07f09b96b23f8b08f2478ef18a12fead1804