Setting enforce_single_del_contracts=false by default
Summary:
RocksDB 7.3 added a strict SingleDelete contract enforcement
that does not allow to mix Delete and SingleDelete for the same key.
As of now, MyRocks Compaction Filter has not complied with the new
contract, even though the contract was not necessary since our Compaction
Filter usage was for purging unused data.
This diff sets enforce_single_del_contracts to false until we fully
follow the strict SingleDelete contract.

@update-submodule: rocksdb

Reviewed By: luqun

Differential Revision: D36566118