Fix multiquery audit logging
Summary:
There is a regression in 8.0 where the query sent to the audit plugin contains the whole request packet, instead of just the query currently being executed.

This means that for multiqueries, the audit plugin will log the same query multiple times.

The fix is to restore what was done in 5.6, and pass actual query length down to the logging code.

Fixes https://bugs.mysql.com/bug.php?id=107390

Reviewed By: hermanlee

Differential Revision: D36648219