Marking the shardbeater as stopped after sb_thd has been destroyed and removed
Summary:
We hit a low frequency race condition in testing, where
running "stop shardbeater; start shardbeater" can cause the server to
crash. This was happening because we set the state of the SHARDBEATER to
STOPPED a bit too early. This unblocked the stop shardbeater SQL command
and allowed it to pick up the start shardbeater SQL next, while the
generate_shardbeats thread was still running and would next destroy
sb_thd and remove it from global threads.

This led to the start shardbeats SQL to start a THD which was
immediately deleted causing the crash.

Reviewed By: abhinav04sharma

Differential Revision: D36879642