add a missing file to libservices
Summary:
Previously, I added a definition for a new service, rpc_plugin_service,
to include/mysql/service_rpc_plugin.h file, but not in libservices.
As a result, if a test plugin calling rpc_plugin_service is built with
MYSQL_DYNAMIC_PLUGIN option, it fails to find it.
This diff fixes this issue by adding rpc_plugin_service to libservices.

Reviewed By: abal147

Differential Revision: D36888689

fbshipit-source-id: e73a86384815d8f3db409896ab25a95d8b07dc70