Skip low limit heuristic if current index already provides order
Summary:
The low limit heuristic is only useful if the current best plan does not provide index order. The code attempts to check for this, but it does not work well for `ORDER BY DESC`.

The problem is that if we detect a a descending scan, we will call `reverse_sorted` on the range plan to see if it provides reverse order. However, at this point in optimization, range plans are always forwards sorted because `test_quick_select` always returns forwards sorted plans, and reversing them is one of the final steps of optimization.

Instead we should check if they are reversible via `reverse_sort_possible`.

I've filed an upstream bug (https://bugs.mysql.com/107626), although it may be low priority for them if it only affects the classic optimizer.

Reviewed By: luqun

Differential Revision: D37321652

fbshipit-source-id: 69282c46dbcd347d23194f15268578b08ab7caac